app_storage: rename source files
